requirements.txt
APScheduler==3.11.0
Authlib==1.3.2
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-APScheduler==1.13.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
elastic-apm
pyyaml